requirements.txt
# requirements.txt

python-escpos